fix: refactor util in teacher mode and add types
