+++ package.json
@@ -47,7 +47,7 @@
     "start:local": "env-cmd -f ./.env.local react-scripts start",
     "start:api": "json-server --watch data/db.json --port 3636 --id _id --routes data/routes.json",
     "prestart:api": "cat data/default.json > data/db.json",
-    "build": "run-s build:react build:bundle",
+    "build": "run-s build:react",
     "build:react": "react-scripts build",
     "build:bundle": "webpack --config webpack.config.js --mode=production",
     "build:local": "env-cmd -f ./.env.local yarn build",
