+++ scripts/deploy.sh
@@ -9,7 +9,7 @@ usage() {
 }
 
 # default build directory
-BUILD="dist/"
+BUILD="build/"
 
 # default version
 REACT_APP_VERSION="latest"
@@ -112,12 +112,12 @@ aws s3 sync ${BUILD} s3://${APP_DIR} --delete
 # see: https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_aws-services-that-work-with-iam.html
 
 # ensure the correct distribution variables are defined
-#if \
-#  [ -z "${DISTRIBUTION}" ]; then
-#  echo "error: environment variable DISTRIBUTION is not defined" 1>&2
-#  echo "error: contact your favourite Graasp engineer if you keep running into trouble" 1>&2
-#  exit 1
-#fi
+if \
+  [ -z "${DISTRIBUTION}" ]; then
+  echo "error: environment variable DISTRIBUTION is not defined" 1>&2
+  echo "error: contact your favourite Graasp engineer if you keep running into trouble" 1>&2
+  exit 1
+fi
 
-# invalidate cloudfront distribution
-# aws cloudfront create-invalidation --distribution-id ${DISTRIBUTION} --paths /${APP_PATH}/*
+ invalidate cloudfront distribution
+ aws cloudfront create-invalidation --distribution-id ${DISTRIBUTION} --paths /${APP_PATH}/*
