+++ src/components/modes/teacher/Widgets/types/index.js
@@ -1,0 +1,4 @@
+export const USER_ID = 'userId';
+export const DATE = 'date';
+export const VERB = 'verb';
+export const CREATED_AT = 'createdAt';
